[pdfjs] use pdf.js web worker
The js file must be loaded dynamically for a web worker. Therefore, we
minify the pdf.worker.js file (if not in debug mode) and place it in the
assets dir. Browserify then only needs to require the standard pdf.js
file instead of pdf.combined.js.

This also reduces the initial loading time on pages without a pdf view
since the 1.2MB js file is loaded only when a pdf is actually displayed.